# openapi.yaml
openapi: 3.0.0
info:
  title: Utdon API Documentation
  version: 1.1.0
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
  responses:
    UnauthorizedError:
      description: Access token,user/password are missing, invalid, or session is not set
  schemas:
    Error:
      type: string
      description: Catched Error as string
...
    ChangePasswordType:
      type: object
      properties:
        password:
          type: string
        newPassword:
          type: string
        newConfirmPassword:
          type: string
    NewUserType:
      type: object
      properties:
        login:
          type: string
          description: login of the new user
          example: AzureDiamond
        password:
          type: string
          description: password of the new user
          example: hunter2
    InfoIuType:
      type: object
      properties:
        login:
          type: string
        bearer:
          type: string
...
security:
  - ApiKeyAuth: []
servers:
  - url: /api/v1
paths:
...
  /userlogin:
    post:
      summary: login to the system with UI
      description: UI login method
      security:
        - ApiKeyAuth: []
      tags:
        - Authentication
      requestBody:
        description: login properties
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                login:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Login OK
          content:
            application/text:
              schema:
                $ref: '#/components/schemas/InfoIuType'
        "401":
          $ref: '#/components/responses/UnauthorizedError'
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users:
    get:
      summary: get users list
      description: Used by the User Management UI to get users list
      security:
        - ApiKeyAuth: []
      tags:
        - Authentication
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        "401":
          $ref: '#/components/responses/UnauthorizedError'
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: create user in the database
      description: UI create user method
      security:
        - ApiKeyAuth: []
      tags:
        - Authentication
      requestBody:
        description: login and password
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewUserType'
      responses:
        "200":
          description: created user
          content:
            application/text:
              schema:
                type: object
                properties:
                  login:
                    type: string
        "400":
          description: User already exists
        "401":
          $ref: '#/components/responses/UnauthorizedError'
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{login}:
    delete:
      summary: delete user from the database
      description: UI delete user method
      security:
        - ApiKeyAuth: []
      tags:
        - Authentication
      parameters:
        - in: path
          name: login
          required: true
          description: login of the user to delete
          schema:
            type: string
            example: user
      responses:
        "200":
          description: deleted
          content:
            application/text:
              schema:
                type: object
                properties:
                  login:
                    type: string
        "401":
          $ref: '#/components/responses/UnauthorizedError'
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /isauthenticated:
    get:
      summary: is user logged
      description: Used by UI to verify user is logged
      security:
        - ApiKeyAuth: []
      tags:
        - Authentication
      responses:
        "200":
          description: User is logged
          content:
            application/json:
              schema:
                type: string
        "401":
          $ref: '#/components/responses/UnauthorizedError'
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /userlogout:
    get:
      summary: user logout method
      description: Used by UI to logout user
      security:
        - ApiKeyAuth: []
      tags:
        - Authentication
      responses:
        "204":
          description: User is logged out
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /changepassword:
    put:
      summary: user change password
      description: Used by UI to change user password
      security:
        - ApiKeyAuth: []
      tags:
        - Authentication
      requestBody:
        description: passwords list
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordType'
      responses:
        "204":
          description: Password has been changed
        "401":
          $ref: '#/components/responses/UnauthorizedError'
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /bearer:
    get:
      summary: get user user auth Token
      description: Used by UI to get user auth Token
      security:
        - ApiKeyAuth: []
      tags:
        - Authentication
      responses:
        "200":
          description: User auth Token
          content:
            application/json:
              schema:
                type: object
                properties:
                  bearer:
                    type: string
        "401":
          $ref: '#/components/responses/UnauthorizedError'
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: change user auth Token
      description: Used by UI to get new user auth Token
      security:
        - ApiKeyAuth: []
      tags:
        - Authentication
      responses:
        "204":
          description: User auth Token has been changed
        "401":
          $ref: '#/components/responses/UnauthorizedError'
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /user:
    get:
      summary: get user's login
      description: Used by UI to show the user's login in the header
      security:
        - ApiKeyAuth: []
      tags:
        - Authentication
      responses:
        "200":
          description: User's login
          content:
            application/json:
              schema:
                type: object
                properties:
                  login:
                    type: string
        "401":
          $ref: '#/components/responses/UnauthorizedError'
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...